add: 상세정보 scroll view  내 content view 파일 분리
